% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01.R
\name{buildDataSetPackage}
\alias{buildDataSetPackage}
\title{Preprocess, document and build a data package}
\usage{
buildDataSetPackage(packageName = NULL, vignettes = FALSE)
}
\arguments{
\item{packageName}{\code{character} path to package source directory.}
}
\description{
Combines the preprocessing, documentation, and build steps into one.
}

