% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataversion.r
\name{dataVersion}
\alias{dataVersion}
\alias{data_version}
\title{Get the DataVersion for a package}
\usage{
dataVersion(pkg, lib.loc = NULL)

data_version(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{\code{character} the package name}

\item{lib.loc}{\code{character} path to library location.}
}
\description{
Retreives the DataVersion of a package if available
}
\note{
\code{dataVersion()} has been renamed to \code{data_version()}
}
\seealso{
\code{\link[utils]{packageVersion}}
}
