% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{datapackage.skeleton}
\alias{datapackage.skeleton}
\title{Create a Data Package skeleton for use with DataPackageR.}
\usage{
datapackage.skeleton(name = "anRpackage", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character(), r_object_names = character())
}
\arguments{
\item{name}{see \code{\link[utils]{package.skeleton}}}

\item{list}{see \code{\link[utils]{package.skeleton}}}

\item{environment}{see \code{\link[utils]{package.skeleton}}}

\item{path}{see \code{\link[utils]{package.skeleton}}}

\item{force}{see \code{\link[utils]{package.skeleton}}}

\item{code_files}{Optional character vector of paths to Rmd files that process raw data
into R objects. Treated differently that \code{code_files} in \code{\link[utils]{package.skeleton}}. 
Will always pass an empty \code{character()} vector to that function.}
}
\description{
Creates a package skeleton for use with preprpocessData. Creates the additional information needed for versioning
datasets, namely the DataVersion string in DESCRIPTION, DATADIGEST, and the data-raw directory. Updates Read-and-delete-me
to reflect the additional necessary steps.
}
\examples{
\dontrun{
datapackage.skeleton(name="MyDataPackage",path="/tmp")
}
}
