% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackageR-defunct.R
\name{DataPackageR-defunct}
\alias{DataPackageR-defunct}
\alias{datapackage.skeleton}
\alias{dataVersion}
\alias{keepDataObjects}
\title{Defunct functions in package \pkg{DataPackageR}.}
\usage{
datapackage.skeleton(...)

dataVersion(...)

keepDataObjects(...)
}
\arguments{
\item{...}{All arguments are now ignored.}
}
\value{
Defunct function. No return value.
}
\description{
These functions are defunct and no longer supported.
Calling them will result in an error.

When possible, alternatives are suggested.
}
