% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{DataPackageR}
\alias{DataPackageR}
\title{Process data generation code in 'data-raw'}
\usage{
DataPackageR(arg = NULL)
}
\arguments{
\item{arg}{\code{character} name of the package to build.}
}
\value{
logical TRUE if succesful, FALSE, if not.
}
\description{
Assumes .R files in 'data-raw' generate rda files to be stored in 'data'.
Sources datasets.R which can source other R files.
R files sourced by datasets.R must invoke \code{sys.source('myRfile.R',env=topenv())}.
Meant to be called before R CMD build.
}
