% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{datapackager_object_read}
\alias{datapackager_object_read}
\title{Read an object created in a previously run processing script.}
\usage{
datapackager_object_read(name)
}
\arguments{
\item{name}{\code{character} the name of the object. Must be a
name available in the configuration objects. Other objects are not saved.}
}
\value{
An R object.
}
\description{
Read an object created in a previously run processing script.
}
