% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/01.R
\name{dataVersion}
\alias{dataVersion}
\title{Get the DataVersion for a package}
\usage{
dataVersion(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{\code{character} the package name}

\item{lib.loc}{\code{character} path to library location.}
}
\description{
Retreives the DataVersion of a package if available
}
\seealso{
\link{utils:packageVersion}
}

