% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{dateiIds}
\alias{dateiIds}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with ids and labels for combinations of datei.}
\usage{
data(dateiIds)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item id. Datei Id
  \item labels. Labels
}
}
\keyword{datasets}

