% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{do_digests}
\alias{do_digests}
\title{do_digests() function extracted out from DataPackageR}
\usage{
do_digests(pkg_dir, dataenv)
}
\arguments{
\item{pkg_dir}{The top level file path of the data package}

\item{dataenv}{The data environment, from DataPackageR}
}
\value{
TRUE if success
}
\description{
do_digests() function extracted out from DataPackageR
}
