% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{do_doc}
\alias{do_doc}
\title{do_doc() function extracted out from end of DataPackageR}
\usage{
do_doc(pkg_dir, dataenv)
}
\arguments{
\item{pkg_dir}{The top level file path of the data package}

\item{dataenv}{The data environment, from DataPackageR}
}
\value{
TRUE if success
}
\description{
do_doc() function extracted out from end of DataPackageR
}
