% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpStructure}
\alias{dpStructure}
\title{Check if a datapackage has same structure as a reference datapackage}
\usage{
dpStructure(l)
}
\arguments{
\item{l}{Datapackage}
}
\value{
json string with the contents of the Datapackage
}
\description{
Check if two datapackages have the same structure
}
\examples{
\dontrun{
}
}

