% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years:Load accidents by Year & Month.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector with list of years}
}
\value{
A data.frame with number of accidents by years summarized by month
}
\description{
Extract the month from the years arg passed to function.
And check if file exists for those months.
If it doesnt exist throw warning "invalid year".
}
\details{
Ancillary function used by \code{fars_summarize_years}
}
