% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years:No.of Accidents per Month/Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with a list of years to summarize by.}
}
\value{
a data.frame with number of accidents by years summarized by month
}
\description{
Retrieve the files based on the year.
group by year and month. Get n records per year .
Summarize No of records based on years and Month.
}
