% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{get_yml_objects}
\alias{get_yml_objects}
\title{Get data objects from YAML configuration}
\usage{
get_yml_objects(ymlconf)
}
\arguments{
\item{ymlconf}{YAML configuration list produced by validate_yml()}
}
\value{
Character vector of object names
}
\description{
Get data objects from YAML configuration
}
