% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{get_yml_r_files}
\alias{get_yml_r_files}
\title{Get R and Rmd files from YAML configuration}
\usage{
get_yml_r_files(ymlconf)
}
\arguments{
\item{ymlconf}{YAML configuration list produced by validate_yml()}
}
\value{
Character vector of enabled R and Rmd files specified in YAML file
}
\description{
Get R and Rmd files from YAML configuration
}
