% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAeDpStr}
\alias{getAeDpStr}
\title{Check if a datapackage has same ae_structure as a reference datapackage}
\usage{
getAeDpStr(dpStr)
}
\arguments{
\item{dp1}{Datapackage}

\item{dp2}{Datapackage}
}
\value{
json string with the contents of the Datapackage
}
\description{
Check if two datapackages have the same structure
}
\examples{
\dontrun{
}
}

