% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDataframe}
\alias{getDataframe}
\title{Get the data contents of a datapkg object as a data frame}
\usage{
getDataframe(dp, dtIdx = 1, withNames = FALSE)
}
\arguments{
\item{dp}{Input datapackage}

\item{dtIdx}{Index of the resource to get the data frame from. Defaults to 1.}
}
\value{
data frame
}
\description{
Get the contents of a data package as a data frame
}
\examples{
\dontrun{
dp <- newDatapkg(mtcars)
getDataframe(dp, withNames=TRUE)
}
}

