% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDataframes}
\alias{getDataframes}
\title{Get all data contents of a datapkg object as a list of data frames.}
\usage{
getDataframes(dp, withNames = FALSE)
}
\arguments{
\item{dp,}{dtIdx is the id of the resource}
}
\value{
data frame
}
\description{
Get all data contents of a datapkg object as a list of data frames.
}
\examples{
\dontrun{
dp <- newDatapkg(list(mt = mtcars[1:4,], mycars = cars[1:2,]))
getDataframes(dp, withNames = TRUE)
}
}

