% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDatatypeByIdx}
\alias{getDatatypeByIdx}
\title{Get datatype by resource index}
\usage{
getDatatypeByIdx(dp, dtIdx = 1)
}
\arguments{
\item{dp}{datapackage}

\item{Idx}{resource index, defaults to 1.}
}
\value{
character vector of datatypes for datapackage resource with index = Idx
}
\description{
Get datatype by resource index
}
\examples{
\dontrun{
}
}

