% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDateiSample}
\alias{getDateiSample}
\title{Creates random datapackages for a given datatype.}
\usage{
getDateiSample(dateiId, asDp = FALSE, random = TRUE)
}
\arguments{
\item{site}{path to the directory you want to set up}
}
\value{
logical TRUE if successful, FALSE if directory already exists
}
\description{
Sets up the directory structure for a new static site
}
\examples{
\dontrun{
mysite <- "/home/david/github/mysite"
skeleton(mysite)
}
}

