% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDpSelectionByName}
\alias{getDpSelectionByName}
\title{Set fieldnames in a list}
\usage{
getDpSelectionByName(dp, dtIdx = 1, cols = c())
}
\arguments{
\item{dp}{datapackage}
}
\value{
dp with updated fieldanmes
}
\description{
Set fieldnames by resource index
}
\examples{
\dontrun{
}
}

