% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFieldNamesByIdx}
\alias{getFieldNames}
\alias{getFieldNamesByIdx}
\alias{setFieldNames}
\title{Get fieldnames by resource index}
\usage{
getFieldNamesByIdx(dp, dtIdx = 1)

getFieldNames(dp, withLabels = FALSE)

setFieldNames(dp, fieldNames)
}
\arguments{
\item{dp}{datapackage}

\item{Idx}{resource index, defaults to 1.}

\item{dp}{datapackage}

\item{dp}{datapackage}
}
\value{
character vector of datatypes for datapackage resource with index = Idx

list of fieldanmes for the datapackage

dp with updated fieldanmes
}
\description{
Get fieldnames by resource index

Get fieldnames by resource index

Set fieldnames by resource index
}
\examples{
\dontrun{
}
\dontrun{
}
\dontrun{
}
}

