% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getResourceNames}
\alias{getResourceNames}
\title{Get datatypes for all resources in a list}
\usage{
getResourceNames(dp)
}
\arguments{
\item{dp}{datapackage}
}
\value{
list of data_types for each resource
}
\description{
Get data_types for all resources in a list
}
\examples{
\dontrun{
}
}

