% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{keepDataObjects-defunct}
\alias{keepDataObjects-defunct}
\alias{keepDataObjects}
\title{These functions are no longer available.}
\usage{
keepDataObjects(...)
}
\arguments{
\item{...}{arguments}
}
\description{
These functions are no longer available.
}
