% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{keepDataObjects}
\alias{keepDataObjects}
\title{Specify which data objects to keep}
\usage{
keepDataObjects(obj)
}
\arguments{
\item{obj}{\code{character} vector of object names}
}
\description{
Specify the names of the data objects to keep. To be called after all preprocessing code.
}
