% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_save.R
\name{load_all_datasets}
\alias{load_all_datasets}
\title{Load all datasets in a package}
\usage{
load_all_datasets(packageName)
}
\arguments{
\item{packageName}{\code{character} the package name}
}
\description{
loads all datasets in a package
}
