% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadDpData}
\alias{loadDpData}
\title{Load data in a data field for each resource}
\usage{
loadDpData(dp, dpPath = ".")
}
\arguments{
\item{Datapkg}{to load the data from}
}
\value{
Datapgk with updated data field
}
\description{
#' Load data in a data field for each resource
}
\examples{
\dontrun{
}
}

