% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename:Print file Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{to load the data for that year.}
}
\value{
a filename

output
"accident_2013.csv.bz2"
}
\description{
Print file name for a particular year.
}
