% Generated by roxygen2: do not edit by hand

\name{matchup-package}
\alias{matchup-paclage}
\title{machup-package}
\usage{match(x= NULL, y=NULL, dp ="TRUE or FALSE" )
}
\arguments{
\item{x}{\code{character} short list vector.}
\item{y}{\code{character} long vector}
\item{dp}{\code{logical} TRUE if you want duplicates}

\description{
Print a vector whit the coincidences.
}
\examples{
matchup<-function(x,y,dp){match(t(x),t(y)) ->z

  z<-z[!is.na(z)]



  y[z]->m

  dp==c(TRUE,FALSE)
  if(dp==TRUE){h <- m[!duplicated(m)]
  h}
  else{m}
}
