% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{package_build}
\alias{package_build}
\title{Preprocess, document and build a data package}
\usage{
package_build(packageName = NULL, vignettes = FALSE, masterfile = NULL)
}
\arguments{
\item{packageName}{\code{character} path to package source directory. Defaults to the current path when NULL.}

\item{vignettes}{\code{logical} specify whether to build vignettes. Default FALSE.}

\item{masterfile}{\code{characer} path to file in data-raw that sources processing scripts. Will do
a partial build of the package.}
}
\description{
Combines the preprocessing, documentation, and build steps into one.
}
