% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processData.R
\docType{package}
\name{preprocessData-package}
\alias{preprocessData-package}
\title{preprocessData}
\description{
Package to install the preprocessData script that can be called via the R CMD mechanism.
R CMD preprocessData packagename looks for .R files in "data-raw" within the "packagename" package source tree.
It sources all the .R files in "data-raw". The .R files are expected to read "raw data" from some source, like
"inst/extdata", or from an online url, process them in some way, such that they are tidy and standardized, and
save the resulting data frames or data.tables in "data".  The user should also document the data sets either
using "roxygen2"  by including an .R file under the "R" directory, or in user-written .Rd files under the "man"
directory. After the data have been generated, the user can call R CMD build packagename.
}

