% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_data_path}
\alias{project_data_path}
\title{Get DataPackageR data path}
\usage{
project_data_path()
}
\value{
\code{character}
}
\description{
Get DataPackageR data path
}
\details{
Returns the path to the datapackage data directory.
}
\examples{
project_data_path()
}
