% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_extdata_path}
\alias{project_extdata_path}
\title{Get DataPackageR extdata path}
\usage{
project_extdata_path(file = NULL)
}
\arguments{
\item{file}{\code{character} or \code{NULL} (default).}
}
\value{
\code{character}
}
\description{
Get DataPackageR extdata path
}
\details{
Returns the path to the data package extdata subdirectory, or
constructs a path to a file in the extdata subdirectory from the
file argument.
}
\examples{
if(rmarkdown::pandoc_available()){
project_extdata_path(file = "mydata.csv")
}
}
