% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_path}
\alias{project_path}
\title{Get DataPackageR Project Root Path}
\usage{
project_path(file = NULL)
}
\arguments{
\item{file}{\code{character} or \code{NULL} (default).}
}
\value{
\code{character}
}
\description{
Get DataPackageR Project Root Path
}
\details{
Returns the path to the data package project root, or
constructs a path to a file in the project root from the
file argument.
}
\examples{
project_path( file = "DESCRIPTION" )
}
