% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readDatapackage}
\alias{readDatapackage}
\title{Read datapackage in json}
\usage{
readDatapackage(dp, dpPath = ".")
}
\arguments{
\item{dp}{}
}
\value{
logical TRUE if successful, FALSE if directory already exists
}
\description{
Write datapackage in json
}
\examples{
\dontrun{
mysite <- "/home/david/github/mysite"
skeleton(mysite)
}
}

