% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readDfString}
\alias{readDfString}
\title{Load data in a data frame}
\usage{
readDfString(pasted)
}
\arguments{
\item{Datapkg}{to load the data from by resource index.}
}
\value{
Datapgk with updated data field
}
\description{
#' Load data in a data frame by reading the paths of the resources
}
\examples{
\dontrun{
}
}

