% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_remote_file.R
\name{retrieve_remote_file}
\alias{retrieve_remote_file}
\title{retrieve_remote_file}
\usage{
retrieve_remote_file(url, user = NULL, password = NULL, ...)
}
\arguments{
\item{url}{Single \code{\link{character}} representing the URL to retrieve.}

\item{user}{Single \code{\link{character}} object used for authentication
where retrieval requires it.}

\item{password}{Single \code{\link{character}} object used for authentication
where retrieval requires it.}

\item{...}{Further parameters handed to \code{\link[httr]{authenticate}}.}
}
\value{
Returns a \code{\link{character}} object representing the path of the
temporary file resulting from the retrieval.
}
\description{
Convenience funtion to retrieve a data file from an URL.
}
\details{
Wraps functionality provided by \pkg{httr}.
}
\examples{
# Load tools
require(magrittr)

# Retrieve a remote file (from Billing et al. (2016). Comprehensive transcriptomic
# and proteomic characterization of human mesenchymal stem cells reveals
# source specific cellular markers. Sci Rep 6, 21507.
# Licensed under the Creative Commons Attribution 4.0 International License.
# http://creativecommons.org/licenses/by/4.0/
## Complicated URL generation to circumvent line length restrictions
tmp_url <- paste0(
  c("http://www.nature.com/article-assets/npg/srep",
  "2016/160209/srep21507/extref/srep21507-s4.xls"),
  collapse = "/")
(tmp_path <- tmp_url \%>\%
  retrieve_remote_file())

# Explore the file
\donttest{
  require(readxl)
  tmp_path \%>\%
    read_excel(skip = 1) \%>\%
    str()
}
}
\seealso{
\code{\link{include_data_file}},
\code{\link{fetch_missing_remote_data}}
}
\author{
Johannes Graumann
}
