% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_integrity.R
\name{test_data_integrity}
\alias{test_data_integrity}
\title{test_data_integrity}
\usage{
test_data_integrity(package_name = NULL)
}
\arguments{
\item{package_name}{Single \code{\link{character}} object, providing the name
of the package for which data integrity is to be tested.}
}
\description{
Uses information in the \code{\link{data_catalogue}} to check the
integrity of data sets shipped.
}
\details{
This function allows for integrity checking of data sets included
via \code{\link[digest]{digest}}, using the cryptographic hash information
stored in \code{\link{data_catalogue}}.
}
\author{
Johannes Graumann
}
