% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_raw_data.R
\name{unpack_raw_data}
\alias{unpack_raw_data}
\title{unpack_raw_data}
\usage{
unpack_raw_data(package_name = NULL, file_names = NULL,
  target_dir = tempdir())
}
\arguments{
\item{package_name}{Single \code{\link{character}} naming the
(\pkg{datapackageR}-managed) \code{R} package the data set(s) to be extracted
reside in.}

\item{file_names}{\code{\link{character}} object representing the names
(\code{\link{basename}}s) of the data sets to be extracted. Defaults to all
data sets listed in \code{\link{data_catalogue}} if set to \code{\link{NULL}}.}

\item{target_dir}{Single \code{\link{character}} defining the target
directory into which the data sets are extracted.}
}
\value{
Returns a named \code{\link{character}} object containing the paths
pointing at the extracted data files.
}
\description{
Convenience function to unpack raw data packaged by
\pkg{datapackageR}.
}
\details{
Using \code{\link[utils]{unzip}}, the function unpacks all data sets
listed in \code{\link{data_catalogue}} (or a user-defined selection thereof)
into \code{target_directory}.
}
\examples{
# Load tools
library(magrittr)

# Create a package with immediate data inclusion
## Define a package root
pkg_root <- tempdir() \%>\%
  file.path("packagetest")
## Create a dummy data file
data.frame(
  x   = 1,
  y   = 1:10,
  fac = sample(LETTERS[1:3], 10, replace = TRUE)) \%>\%
write.table(
  file      = file.path(dirname(pkg_root), "data_dummy.tsv"),
  sep       = "\\t",
  col.names = TRUE,
  row.names = FALSE)
## Create the package infrastructure, including the dummy file
init(
  root = pkg_root,
  files_to_include = file.path(dirname(pkg_root), "data_dummy.tsv"),
  file_reading_function = "read.csv",
  file_reading_options = list(sep = "\\t", stringsAsFactors = FALSE))

# Install the resulting package
devtools::install(pkg_root)

# Extract and explore
unpack_raw_data(package_name = "packagetest") \%>\%
  file.info()

# Uninstall the package
remove.packages("packagetest")

# Clean up the package roots - ensure proper example testing by R CMD check
unlink(pkg_root, recursive = TRUE)
}
\seealso{
\code{\link[utils]{unzip}}, \code{\link{data_catalogue}}
}
\author{
Johannes Graumann
}
