% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_data_object}
\alias{use_data_object}
\title{Add a data object to a data package.}
\usage{
use_data_object(object_name = NULL)
}
\arguments{
\item{object_name}{Name of the data object. Should be created by a processing script in data-raw. \code{character} vector of length 1.}
}
\value{
invisibly returns TRUE for success.
}
\description{
The data object will be added to the yml configuration file.
}
\examples{
if(rmarkdown::pandoc_available()){
myfile <- tempfile()
file <- system.file("extdata", "tests", "extra.rmd",
                     package = "DataPackageR")
datapackage_skeleton(
  name = "datatest",
  path = tempdir(),
  code_files = file,
  force = TRUE,
  r_object_names = "data")
use_data_object(object_name = "newobject")
}

}
