% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{use_ignore}
\alias{use_ignore}
\title{Ignore specific files by git and R build.}
\usage{
use_ignore(file = NULL, path = NULL)
}
\arguments{
\item{file}{\code{character} File to ignore.}

\item{path}{\code{character} Path to the file.}
}
\value{
invisibly returns 0.
}
\description{
Ignore specific files by git and R build.
}
\examples{
use_ignore("foo", ".")
}
