% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{validate_package_skeleton}
\alias{validate_package_skeleton}
\title{Validate data package skeleton, extracted out from big DataPackageR function}
\usage{
validate_package_skeleton(pkg_dir)
}
\arguments{
\item{pkg_dir}{Path of top level of data package}
}
\value{
Silently returns pkg_dir if no errors thrown
}
\description{
Validate data package skeleton, extracted out from big DataPackageR function
}
