% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{validate_pkg_name}
\alias{validate_pkg_name}
\title{Check that pkg name inferred from pkg path is same as pkg name in DESCRIPTION}
\usage{
validate_pkg_name(package_path)
}
\arguments{
\item{package_path}{Package path}
}
\value{
Package name (character) if validated
}
\description{
Check that pkg name inferred from pkg path is same as pkg name in DESCRIPTION
}
