% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{validate_yml}
\alias{validate_yml}
\title{Validate YAML file, extracted out from big DataPackageR function}
\usage{
validate_yml(pkg_dir)
}
\arguments{
\item{pkg_dir}{Path of top level of data package}
}
\value{
List object from read_yaml(ymlfile)
}
\description{
Validate YAML file, extracted out from big DataPackageR function
}
