% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WriteDatapackage}
\alias{WriteDatapackage}
\alias{writeDatapackage}
\title{Write datapackage in json}
\usage{
writeDatapackage(dp, path = ".")
}
\arguments{
\item{dp}{}
}
\value{
logical TRUE if successful, FALSE if directory already exists
}
\description{
Write datapackage in json
}
\examples{
\dontrun{
mysite <- "/home/david/github/mysite"
skeleton(mysite)
}
}

