% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{buildDataSetPackage}
\alias{buildDataSetPackage}
\title{Preprocess, document and build a data package}
\usage{
buildDataSetPackage(packageName = NULL, vignettes = FALSE, outpath = NULL,
  masterfile = NULL)
}
\arguments{
\item{packageName}{\code{character} path to package source directory. Defaults to the current path when NULL.}

\item{vignettes}{\code{logical} specify whether to build vignettes. Default FALSE.}

\item{masterfile}{\code{characer} path to file in data-raw that sources processing scripts. Will do 
a partial build of the package.}
}
\description{
Combines the preprocessing, documentation, and build steps into one.
}
