% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataversion.r
\name{data_version}
\alias{data_version}
\alias{dataVersion}
\title{Get the DataVersion for a package}
\usage{
data_version(pkg, lib.loc = NULL)

dataVersion(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{\code{character} the package name}

\item{lib.loc}{\code{character} path to library location.}
}
\description{
Retreives the DataVersion of a package if available
}
\note{
\code{dataVersion()} has been renamed to \code{data_version()}
}
\examples{

f <- tempdir()
f <- file.path(f,"foo.Rmd")
con <- file(f)
writeLines("```{r}\\n a= 100 \\n```\\n",con=con)
close(con)
unlink(file.path(tempdir(),"MyDataPackage"),force = TRUE, recursive = TRUE)
datapackage_skeleton(name="MyDataPackage",
   path=tempdir(), 
   force = TRUE,
   r_object_names = "a",
   code_files = f)
   package_build(file.path(tempdir(),"MyDataPackage"))

   devtools::load_all(file.path(tempdir(),"MyDataPackage"))
   data_version("MyDataPackage")
}
\seealso{
\code{\link[utils]{packageVersion}}
}
