% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.R
\name{datapackage.skeleton}
\alias{datapackage.skeleton}
\title{Create a Data Package skeleton for use with preprocessData.}
\usage{
datapackage.skeleton(name = "anRpackage", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character())
}
\arguments{
\item{name}{\code{character} see \code{\link{utils:package.skeleton}}}

\item{list}{see \code{\link{utils:package.skeleton}}}

\item{environment}{see \code{\link{utils:package.skeleton}}}

\item{path}{see \code{\link{utils:package.skeleton}}}

\item{force}{see \code{\link{utils:package.skeleton}}}

\item{code_files}{see \code{\link{utils:package.skeleton}}}
}
\description{
Creates a package skeleton for use with preprpocessData. Creates the additional information needed for versioning
datasets, namely the DataVersion string in DESCRIPTION, DATADIGEST, and the data-raw directory. Updates Read-and-delete-me
to reflect the additional necessary steps.
}
\examples{
data.package.skeleton(name="MyDataPackage",path="/tmp")
}

