% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{package_build}
\alias{package_build}
\title{Preprocess, document and build a data package}
\usage{
package_build(packageName = NULL, vignettes = FALSE, log = INFO)
}
\arguments{
\item{packageName}{\code{character} path to package source directory. Defaults to the current path when NULL.}

\item{vignettes}{\code{logical} specify whether to build vignettes. Default FALSE.}

\item{log}{log level \code{INFO,WARN,DEBUG,FATAL}}
}
\description{
Combines the preprocessing, documentation, and build steps into one.
}
\examples{

f <- tempdir()
f <- file.path(f,"foo.Rmd")
con <- file(f)
writeLines("```{r}\\na= 100\\n```\\n",con=con)
close(con)
unlink(file.path(tempdir(),"MyDataPackage"),force = TRUE, recursive = TRUE)
datapackage_skeleton(name="MyDataPackage",
   path=tempdir(), 
   force = TRUE,
   r_object_names = "a",
   code_files = f)
   
package_build(file.path(tempdir(),"MyDataPackage"))
}
