% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processData.R
\name{preprocessData}
\alias{preprocessData}
\title{Process data generation code in "data-raw"}
\usage{
preprocessData(arg = NULL)
}
\description{
Assumes .R files in "data-raw" generate rda files to be stored in "data".
Sources each of these .R files.
Meant to be called before R CMD build.
}

