% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_data_path}
\alias{project_data_path}
\title{Get DataPackageR data path}
\usage{
project_data_path(file = NULL)
}
\arguments{
\item{file}{\code{character} or \code{NULL} (default).}
}
\value{
\code{character}
}
\description{
Get DataPackageR data path
}
\details{
Returns the path to the data package data subdirectory, or
constructs a path to a file in the data subdirectory from the
file argument.
}
\examples{
project_data_path( file = "data.rda" )
}
