% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_extdata_path}
\alias{project_extdata_path}
\title{Get DataPackageR extdata path}
\usage{
project_extdata_path()
}
\value{
\code{character}
}
\description{
Get DataPackageR extdata path
}
\details{
Returns the path to the datapackage extdata directory.
}
\examples{
project_extdata_path()
}
