% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{project_path}
\alias{project_path}
\title{Get DataPackageR Project Root Path}
\usage{
project_path()
}
\value{
\code{character}
}
\description{
Get DataPackageR Project Root Path
}
\details{
Returns the path to the data package project root.
}
\examples{
project_path()
}
