% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_raw_dataset}
\alias{use_raw_dataset}
\title{Add a raw data set to data-raw}
\usage{
use_raw_dataset(path = NULL)
}
\arguments{
\item{path}{\code{character} path to file or directory.}
}
\value{
invisibly returns TRUE for success. Stops on failure.
}
\description{
The file or directory specified by \code{path} will be moved into
the data-raw directory.
}
